integration: Move K8sCommand to its own file
